feat: add layout and abstract applying font
