+++ pages/_app.tsx
@@ -1,27 +1,12 @@
 import '@styles/globals.css';
 import type { AppProps } from 'next/app';
-import localFont from 'next/font/local';
 
-const pretendard = localFont({
-  src: [
-    {
-      path: '../public/fonts/Pretendard-Regular.woff2',
-      weight: '400',
-      style: 'normal',
-    },
-    {
-      path: '../public/fonts/Pretendard-Medium.woff2',
-      weight: '500',
-      style: 'normal',
-    },
-  ],
-  variable: '--font-pretendard',
-});
+import { GlobalLayout } from '../shared/components/layout';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
-    <main className={`${pretendard.variable} font-sans`}>
+    <GlobalLayout>
       <Component {...pageProps} />
-    </main>
+    </GlobalLayout>
   );
 }
