+++ shared/components/layout/index.ts
@@ -1,0 +1,3 @@
+import GlobalLayout from './GlobalLayout';
+
+export { GlobalLayout };
